feat : CollectedStamp 및 Stamp 엔터티 수정 (#68)
